requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via httpcore
certifi==2022.9.24
    # via
    #   httpcore
    #   httpx
h11==0.12.0
    # via httpcore
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   rfc3986
python-dotenv==0.21.0
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
